RA-1174 - Split tests around wages out into separate test classes.
